+++ modules/bars/micromenu.lua
@@ -16,21 +16,22 @@ local unpack = _G.unpack
 local C_Timer = _G.C_Timer
 
 --[[ luacheck: globals
-	BreakUpLargeNumbers ContainerIDToInventoryID CreateFrame CursorCanGoInSlot CursorHasItem GameTooltip GetAddOnInfo
-	GetAddOnMemoryUsage GetBagSlotFlag GetContainerNumFreeSlots GetContainerNumSlots GetCurrencyInfo
-	GetInventoryItemDurability GetInventoryItemTexture GetLFGDungeonShortageRewardInfo GetLFGRandomDungeonInfo
-	GetLFGRoles GetLFGRoleShortageRewards GetMoney GetMoneyString GetNetStats GetNumAddOns GetNumRandomDungeons
-	GetNumRFDungeons GetNumSavedInstances GetNumSavedWorldBosses GetQuestResetTime GetRFDungeonInfo GetSavedInstanceInfo
-	GetSavedWorldBossInfo GetTime GuildMicroButtonTabard InCombatLockdown IsAddOnLoaded IsInventoryItemLocked
-	IsInventoryItemProfessionBag IsLFGDungeonJoinable IsShiftKeyDown LFDMicroButton LibStub LSBagBar MainMenuBarDownload
+	AchievementMicroButton_Update BreakUpLargeNumbers ContainerIDToInventoryID CreateFrame CursorCanGoInSlot
+	CursorHasItem GameTooltip GetAddOnInfo GetAddOnMemoryUsage GetBagSlotFlag GetContainerNumFreeSlots
+	GetContainerNumSlots GetCurrencyInfo GetInventoryItemDurability GetInventoryItemTexture
+	GetLFGDungeonShortageRewardInfo GetLFGRandomDungeonInfo GetLFGRoles GetLFGRoleShortageRewards GetMoney
+	GetMoneyString GetNetStats GetNumAddOns GetNumRandomDungeons GetNumRFDungeons GetNumSavedInstances
+	GetNumSavedWorldBosses GetQuestResetTime GetRFDungeonInfo GetSavedInstanceInfo GetSavedWorldBossInfo GetTime
+	GuildMicroButtonTabard InCombatLockdown IsAddOnLoaded IsInventoryItemLocked IsInventoryItemProfessionBag
+	IsKioskModeEnabled IsLFGDungeonJoinable IsShiftKeyDown LFDMicroButton LibStub LSBagBar MainMenuBarDownload
 	MainMenuBarPerformanceBar MicroButtonAndBagsBar MicroButtonPortrait MicroButtonTooltipText PickupBagFromSlot
 	PlaySound PutItemInBag RequestLFDPartyLockInfo RequestLFDPlayerLockInfo RequestRaidInfo SecondsToTime SetBagSlotFlag
 	SetClampedTextureRotation ToggleAllBags UIParent UpdateAddOnMemoryUsage
 
 	BACKPACK_CONTAINER BAG_FILTER_ASSIGN_TO BAG_FILTER_CLEANUP BAG_FILTER_ICONS BAG_FILTER_IGNORE BAG_FILTER_LABELS
-	EQUIP_CONTAINER LE_BAG_FILTER_FLAG_EQUIPMENT LE_BAG_FILTER_FLAG_IGNORE_CLEANUP LE_BAG_FILTER_FLAG_JUNK
-	LFG_ROLE_NUM_SHORTAGE_TYPES NEWBIE_TOOLTIP_SPELLBOOK NUM_BAG_SLOTS NUM_LE_BAG_FILTER_FLAGS
-	PERFORMANCEBAR_MEDIUM_LATENCY SHOW_NEWBIE_TIPS
+	DUNGEONS_BUTTON EQUIP_CONTAINER LE_BAG_FILTER_FLAG_EQUIPMENT LE_BAG_FILTER_FLAG_IGNORE_CLEANUP
+	LE_BAG_FILTER_FLAG_JUNK LFG_ROLE_NUM_SHORTAGE_TYPES NEWBIE_TOOLTIP_LFGPARENT NEWBIE_TOOLTIP_SPELLBOOK NUM_BAG_SLOTS
+	NUM_LE_BAG_FILTER_FLAGS PERFORMANCEBAR_MEDIUM_LATENCY SHOW_NEWBIE_TIPS
 ]]
 
 -- Mine
@@ -989,7 +990,17 @@ do
 
 	function lfdButton_OnEvent(self, event)
 		if event == "LFG_LOCK_INFO_RECEIVED" then
-			self:UpdateIndicator()
+			if GetTime() - (self.lastUpdate or 0) > 9 then
+				self:UpdateIndicator()
+				self.lastUpdate = GetTime()
+			end
+		else
+			if IsKioskModeEnabled() then
+				return
+			end
+
+			self.tooltipText = MicroButtonTooltipText(DUNGEONS_BUTTON, "TOGGLEGROUPFINDER")
+			self.newbieText = NEWBIE_TOOLTIP_LFGPARENT
 		end
 	end
 
@@ -999,7 +1010,7 @@ do
 		if self._config.enabled and self._config.tooltip then
 			self:SetScript("OnEnter", lfdButton_OnEnter)
 
-			self.Ticker = C_Timer.NewTicker(10, function()
+			self.Ticker = C_Timer.NewTicker(15, function()
 				RequestLFDPlayerLockInfo()
 				RequestLFDPartyLockInfo()
 			end)
@@ -1478,7 +1489,10 @@ function MODULE:CreateMicroMenu()
 
 				button.Update = guildButton_Update
 			elseif id == "lfd" then
-				button:HookScript("OnEvent", lfdButton_OnEvent)
+				button:SetScript("OnEvent", lfdButton_OnEvent)
+
+				button.tooltipText = MicroButtonTooltipText(DUNGEONS_BUTTON, "TOGGLEGROUPFINDER")
+				button.newbieText = NEWBIE_TOOLTIP_LFGPARENT
 
 				button.Update = lfdButton_Update
 				button.UpdateIndicator = lfdButton_UpdateIndicator
@@ -1531,6 +1545,11 @@ function MODULE:CreateMicroMenu()
 				repositionAlert(_G[name])
 			end
 		end)
+
+		-- this method was removed, but is still called by the Blizz UI
+		if not AchievementMicroButton_Update then
+			AchievementMicroButton_Update = E.NOOP
+		end
 
 		isInit = true
 	end